// src/cleanup-code.enum.ts
export enum CleanupCode {
    NO_CITAS_DISPONIBLE = 'no_citas_disponible',
    NO_CITAS_DISPONIBLE_ERR = 'no_citas_disponible_err',
    TOO_MANY_REQUESTS = 'too_many_requests',
    TOO_MANY_REQUESTS_ERR = 'too_many_requests_err',
    URL_REJECTED = 'url_rejected',
    URL_REJECTED_ERR = 'url_rejected_err',
    PAGE_1_SEL_OP_ERR = 'page_1_sel_op_err',
    PAGE_1_CLICK_ACC_ERR = 'page_1_click_acc_err',
    PAGE_1_5_CLICK_CONF_ERR = 'page_1_5_click_conf_err',
    PAGE_2_ERR = 'page_2_err',
    PAGE_2_CLICK_ENVIAR_ERR = 'page_2_click_enviar_err',
    PAGE_3_CLICK_ENVIAR_ERR = 'page_3_click_enviar_err',
    PAGE_4_NAVI_ERR = 'page_4_navi_err'
}